// packages/sdk/src/types/parameters/moulinette-list.parameter.type.ts
/**
 * Copyright (c) Minimouli
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */

interface MoulinetteListParameter {
    isOfficial: boolean
    projectName: string
    projectCycle: number
    organizationName: string
}

export type {
    MoulinetteListParameter
}